+++ docs/tutorial/src/sql/comment_add.sql
@@ -1,0 +1,2 @@
+INSERT INTO comment (article_id, body, posted_at)
+VALUES (:articleId, :body, :postedAt);